requirements.txt
dj-database-url==1.2.0
Django==4.1.6
django-admin-thumbnails==0.2.6
django-cors-headers==3.13.0
djangorestframework==3.14.0
Pillow==9.2.0
psycopg2-binary==2.9.5
pycparser==2.21
python-decouple==3.6
requests==2.28.1
whitenoise==6.2.0
